details_view: Place larger ui sections into dedicated functions
