+++ src/components/details_view_components/mod.rs
@@ -1,0 +1,3 @@
+mod log_grid;
+
+pub use log_grid::draw_information_grid;
